--- src/views/MyPage2.vue
@@ -1,1049 +1,0 @@
-<template>
-  <div class="div">
-    <div class="div-2">
-      <div class="div-3">
-        <img
-          loading="lazy"
-          srcSet="https://cdn.builder.io/api/v1/image/assets/TEMP/c74865b6c7feeb69956b0e2e3b237755445eb4a67fceae6585226bda7e04df68?apiKey=d213af4abfe044f4b8c2d54b9aa44ab0&width=100 100w, https://cdn.builder.io/api/v1/image/assets/TEMP/c74865b6c7feeb69956b0e2e3b237755445eb4a67fceae6585226bda7e04df68?apiKey=d213af4abfe044f4b8c2d54b9aa44ab0&width=200 200w, https://cdn.builder.io/api/v1/image/assets/TEMP/c74865b6c7feeb69956b0e2e3b237755445eb4a67fceae6585226bda7e04df68?apiKey=d213af4abfe044f4b8c2d54b9aa44ab0&width=400 400w, https://cdn.builder.io/api/v1/image/assets/TEMP/c74865b6c7feeb69956b0e2e3b237755445eb4a67fceae6585226bda7e04df68?apiKey=d213af4abfe044f4b8c2d54b9aa44ab0&width=800 800w, https://cdn.builder.io/api/v1/image/assets/TEMP/c74865b6c7feeb69956b0e2e3b237755445eb4a67fceae6585226bda7e04df68?apiKey=d213af4abfe044f4b8c2d54b9aa44ab0&width=1200 1200w, https://cdn.builder.io/api/v1/image/assets/TEMP/c74865b6c7feeb69956b0e2e3b237755445eb4a67fceae6585226bda7e04df68?apiKey=d213af4abfe044f4b8c2d54b9aa44ab0&width=1600 1600w, https://cdn.builder.io/api/v1/image/assets/TEMP/c74865b6c7feeb69956b0e2e3b237755445eb4a67fceae6585226bda7e04df68?apiKey=d213af4abfe044f4b8c2d54b9aa44ab0&width=2000 2000w, https://cdn.builder.io/api/v1/image/assets/TEMP/c74865b6c7feeb69956b0e2e3b237755445eb4a67fceae6585226bda7e04df68?apiKey=d213af4abfe044f4b8c2d54b9aa44ab0&"
-          class="img"
-        />
-        <div class="div-4">
-          <span
-            style="
-              font-family:
-                Noto Sans,
-                -apple-system,
-                Roboto,
-                Helvetica,
-                sans-serif;
-              font-weight: 700;
-            "
-            >루챌</span
-          ><span
-            style="
-              font-family:
-                Noto Sans,
-                -apple-system,
-                Roboto,
-                Helvetica,
-                sans-serif;
-            "
-          >
-            챌린저의 하루를 응원합니다!</span
-          >
-        </div>
-      </div>
-      <div class="div-5">
-        <div class="div-6">ABOUT</div>
-        <div class="div-7">CHALLENGE</div>
-        <div class="div-8">LOGOUT</div>
-      </div>
-    </div>
-    <div class="div-9">
-      <div class="div-10">
-        <div class="div-11">
-          <img
-            loading="lazy"
-            srcSet="https://cdn.builder.io/api/v1/image/assets/TEMP/59419233e38974f952407ef71a4fd19d9300e0a71b48bd3920cd59eb58db0e66?apiKey=d213af4abfe044f4b8c2d54b9aa44ab0&width=100 100w, https://cdn.builder.io/api/v1/image/assets/TEMP/59419233e38974f952407ef71a4fd19d9300e0a71b48bd3920cd59eb58db0e66?apiKey=d213af4abfe044f4b8c2d54b9aa44ab0&width=200 200w, https://cdn.builder.io/api/v1/image/assets/TEMP/59419233e38974f952407ef71a4fd19d9300e0a71b48bd3920cd59eb58db0e66?apiKey=d213af4abfe044f4b8c2d54b9aa44ab0&width=400 400w, https://cdn.builder.io/api/v1/image/assets/TEMP/59419233e38974f952407ef71a4fd19d9300e0a71b48bd3920cd59eb58db0e66?apiKey=d213af4abfe044f4b8c2d54b9aa44ab0&width=800 800w, https://cdn.builder.io/api/v1/image/assets/TEMP/59419233e38974f952407ef71a4fd19d9300e0a71b48bd3920cd59eb58db0e66?apiKey=d213af4abfe044f4b8c2d54b9aa44ab0&width=1200 1200w, https://cdn.builder.io/api/v1/image/assets/TEMP/59419233e38974f952407ef71a4fd19d9300e0a71b48bd3920cd59eb58db0e66?apiKey=d213af4abfe044f4b8c2d54b9aa44ab0&width=1600 1600w, https://cdn.builder.io/api/v1/image/assets/TEMP/59419233e38974f952407ef71a4fd19d9300e0a71b48bd3920cd59eb58db0e66?apiKey=d213af4abfe044f4b8c2d54b9aa44ab0&width=2000 2000w, https://cdn.builder.io/api/v1/image/assets/TEMP/59419233e38974f952407ef71a4fd19d9300e0a71b48bd3920cd59eb58db0e66?apiKey=d213af4abfe044f4b8c2d54b9aa44ab0&"
-            class="img-2"
-          />
-          <div class="div-12">루챌</div>
-          <div class="div-13">example@naver.com</div>
-          <div class="div-14">참여중인 챌린지</div>
-          <div class="div-15">개인정보 수정</div>
-        </div>
-        <div class="div-16"></div>
-      </div>
-      <div class="div-17">
-        <div class="div-18">
-          <div class="div-19">
-            <div class="column">
-              <div class="div-20">
-                <div class="div-21">이름</div>
-                <div class="div-22">아이디</div>
-              </div>
-            </div>
-            <div class="column-2">
-              <div class="div-23">
-                <div class="div-24">개인정보 수정</div>
-                <div class="div-25">성이름</div>
-                <div class="div-26">idchallenger00</div>
-              </div>
-            </div>
-          </div>
-        </div>
-        <div class="div-27">
-          <div class="div-28">현재 비밀번호</div>
-          <div class="div-29">현재 비밀번호<span style="color: rgba(255, 0, 0, 1)"> *</span></div>
-        </div>
-        <div class="div-30">
-          <div class="div-31">비밀번호 변경</div>
-          <div class="div-32">새로운 비밀번호</div>
-        </div>
-        <div class="div-33">새로운 비밀번호 확인</div>
-        <div class="div-34">
-          <div class="div-35">
-            <div class="div-36">닉네임</div>
-            <div class="div-37">프로필 사진</div>
-            <div class="div-38">생년월일</div>
-            <div class="div-39">이메일</div>
-          </div>
-          <div class="div-40">
-            <div class="div-41">루챌</div>
-            <div class="div-42">
-              <div class="div-43">
-                <div class="column-3">
-                  <img
-                    loading="lazy"
-                    srcSet="https://cdn.builder.io/api/v1/image/assets/TEMP/b85d43463f67bc07bea2940463d037d560821ccbd3179d2a40ea066ac106ba29?apiKey=d213af4abfe044f4b8c2d54b9aa44ab0&width=100 100w, https://cdn.builder.io/api/v1/image/assets/TEMP/b85d43463f67bc07bea2940463d037d560821ccbd3179d2a40ea066ac106ba29?apiKey=d213af4abfe044f4b8c2d54b9aa44ab0&width=200 200w, https://cdn.builder.io/api/v1/image/assets/TEMP/b85d43463f67bc07bea2940463d037d560821ccbd3179d2a40ea066ac106ba29?apiKey=d213af4abfe044f4b8c2d54b9aa44ab0&width=400 400w, https://cdn.builder.io/api/v1/image/assets/TEMP/b85d43463f67bc07bea2940463d037d560821ccbd3179d2a40ea066ac106ba29?apiKey=d213af4abfe044f4b8c2d54b9aa44ab0&width=800 800w, https://cdn.builder.io/api/v1/image/assets/TEMP/b85d43463f67bc07bea2940463d037d560821ccbd3179d2a40ea066ac106ba29?apiKey=d213af4abfe044f4b8c2d54b9aa44ab0&width=1200 1200w, https://cdn.builder.io/api/v1/image/assets/TEMP/b85d43463f67bc07bea2940463d037d560821ccbd3179d2a40ea066ac106ba29?apiKey=d213af4abfe044f4b8c2d54b9aa44ab0&width=1600 1600w, https://cdn.builder.io/api/v1/image/assets/TEMP/b85d43463f67bc07bea2940463d037d560821ccbd3179d2a40ea066ac106ba29?apiKey=d213af4abfe044f4b8c2d54b9aa44ab0&width=2000 2000w, https://cdn.builder.io/api/v1/image/assets/TEMP/b85d43463f67bc07bea2940463d037d560821ccbd3179d2a40ea066ac106ba29?apiKey=d213af4abfe044f4b8c2d54b9aa44ab0&"
-                    class="img-3"
-                  />
-                </div>
-                <div class="column-4"><div class="div-44">업로드</div></div>
-                <div class="column-5">
-                  <div class="div-45">선택된 파일 없음</div>
-                </div>
-              </div>
-            </div>
-            <div class="div-46">
-              <div class="div-47">2000/01/01</div>
-              <img
-                loading="lazy"
-                src="https://cdn.builder.io/api/v1/image/assets/TEMP/a7ceeef0f86ed5ea00a5596f4479a86ffcf22f5be9f758816bd62371ddab9c7d?apiKey=d213af4abfe044f4b8c2d54b9aa44ab0&"
-                class="img-4"
-              />
-            </div>
-            <div class="div-48">
-              <div class="div-49">example</div>
-              <div class="div-50">@</div>
-              <div class="div-51">
-                <div class="div-52">naver.com</div>
-                <div class="div-53">v</div>
-              </div>
-            </div>
-          </div>
-        </div>
-        <div class="div-54">
-          <div class="div-55">관심 챌린지</div>
-          <div class="div-56">
-            <div class="div-57"></div>
-            <div class="div-58">자기계발</div>
-            <div class="div-59">v</div>
-            <div class="div-60">학습</div>
-            <div class="div-61"></div>
-            <div class="div-62">운동</div>
-            <div class="div-63"></div>
-            <div class="div-64">건강</div>
-            <div class="div-65"></div>
-            <div class="div-66">기타</div>
-          </div>
-        </div>
-        <div class="div-67">수정</div>
-        <div class="div-68">회원탈퇴</div>
-      </div>
-    </div>
-  </div>
-</template>
-
-<style scoped>
-.div {
-  background-color: #fff;
-  display: flex;
-  flex-direction: column;
-}
-.div-2 {
-  background-color: rgba(255, 153, 0, 0.07);
-  display: flex;
-  width: 100%;
-  justify-content: space-between;
-  gap: 20px;
-  white-space: nowrap;
-  padding: 20px 70px;
-}
-@media (max-width: 991px) {
-  .div-2 {
-    max-width: 100%;
-    flex-wrap: wrap;
-    white-space: initial;
-    padding: 0 20px;
-  }
-}
-.div-3 {
-  display: flex;
-  justify-content: space-between;
-  gap: 20px;
-  font-size: 24px;
-  color: #333;
-  font-weight: 400;
-  line-height: 30px;
-}
-@media (max-width: 991px) {
-  .div-3 {
-    max-width: 100%;
-    flex-wrap: wrap;
-    white-space: initial;
-  }
-}
-.img {
-  aspect-ratio: 0.94;
-  object-fit: auto;
-  object-position: center;
-  width: 52px;
-}
-.div-4 {
-  font-family: Inter, sans-serif;
-  flex-grow: 1;
-  margin: auto 0;
-}
-@media (max-width: 991px) {
-  .div-4 {
-    white-space: initial;
-  }
-}
-.div-5 {
-  align-self: end;
-  display: flex;
-  margin-top: 34px;
-  justify-content: space-between;
-  gap: 20px;
-  font-size: 16px;
-  color: #999;
-  font-weight: 700;
-  text-align: center;
-  line-height: 187.5%;
-}
-@media (max-width: 991px) {
-  .div-5 {
-    white-space: initial;
-  }
-}
-.div-6 {
-  font-family:
-    Noto Sans,
-    sans-serif;
-}
-.div-7 {
-  font-family:
-    Noto Sans,
-    sans-serif;
-}
-.div-8 {
-  font-family:
-    Noto Sans,
-    sans-serif;
-}
-.div-9 {
-  align-self: center;
-  display: flex;
-  margin-top: 15px;
-  width: 100%;
-  max-width: 1392px;
-  justify-content: space-between;
-  gap: 20px;
-}
-@media (max-width: 991px) {
-  .div-9 {
-    max-width: 100%;
-    flex-wrap: wrap;
-  }
-}
-.div-10 {
-  display: flex;
-  justify-content: space-between;
-  gap: 20px;
-  font-size: 26px;
-  color: #333;
-  font-weight: 700;
-  text-align: center;
-  line-height: 192%;
-}
-.div-11 {
-  align-self: start;
-  display: flex;
-  margin-top: 50px;
-  flex-direction: column;
-  align-items: center;
-}
-@media (max-width: 991px) {
-  .div-11 {
-    margin-top: 40px;
-  }
-}
-.img-2 {
-  aspect-ratio: 1;
-  object-fit: auto;
-  object-position: center;
-  width: 200px;
-}
-.div-12 {
-  font-family:
-    Noto Sans,
-    sans-serif;
-  line-height: 115%;
-  margin-top: 28px;
-}
-.div-13 {
-  align-self: stretch;
-  margin-top: 11px;
-  font:
-    400 20px/100% Noto Sans,
-    sans-serif;
-}
-.div-14 {
-  color: #555;
-  font-family:
-    Noto Sans,
-    sans-serif;
-  font-weight: 400;
-  margin-top: 110px;
-}
-@media (max-width: 991px) {
-  .div-14 {
-    margin-top: 40px;
-  }
-}
-.div-15 {
-  color: #555;
-  font-family:
-    Noto Sans,
-    sans-serif;
-  margin-top: 46px;
-}
-@media (max-width: 991px) {
-  .div-15 {
-    margin-top: 40px;
-  }
-}
-.div-16 {
-  border-right: 3px solid rgba(255, 153, 0, 0.07);
-  width: 3px;
-  height: 1295px;
-}
-.div-17 {
-  display: flex;
-  flex-direction: column;
-  margin: auto 0;
-}
-@media (max-width: 991px) {
-  .div-17 {
-    max-width: 100%;
-  }
-}
-.div-18 {
-  width: 441px;
-  max-width: 100%;
-}
-.div-19 {
-  gap: 20px;
-  display: flex;
-}
-@media (max-width: 991px) {
-  .div-19 {
-    flex-direction: column;
-    align-items: stretch;
-    gap: 0px;
-  }
-}
-.column {
-  display: flex;
-  flex-direction: column;
-  line-height: normal;
-  width: 29%;
-  margin-left: 0px;
-}
-@media (max-width: 991px) {
-  .column {
-    width: 100%;
-  }
-}
-.div-20 {
-  display: flex;
-  margin-top: 90px;
-  flex-grow: 1;
-  flex-direction: column;
-  font-size: 20px;
-  color: #333;
-  font-weight: 700;
-  white-space: nowrap;
-  line-height: 275%;
-}
-@media (max-width: 991px) {
-  .div-20 {
-    margin-top: 40px;
-    white-space: initial;
-  }
-}
-.div-21 {
-  font-family:
-    Noto Sans,
-    sans-serif;
-}
-.div-22 {
-  font-family:
-    Noto Sans,
-    sans-serif;
-  margin-top: 66px;
-}
-@media (max-width: 991px) {
-  .div-22 {
-    margin-top: 40px;
-  }
-}
-.column-2 {
-  display: flex;
-  flex-direction: column;
-  line-height: normal;
-  width: 71%;
-  margin-left: 20px;
-}
-@media (max-width: 991px) {
-  .column-2 {
-    width: 100%;
-  }
-}
-.div-23 {
-  display: flex;
-  flex-grow: 1;
-  flex-direction: column;
-  font-size: 20px;
-  color: #000;
-  font-weight: 400;
-  line-height: 275%;
-}
-@media (max-width: 991px) {
-  .div-23 {
-    margin-top: 40px;
-  }
-}
-.div-24 {
-  color: #ec8e00;
-  text-align: center;
-  align-self: end;
-  font:
-    700 28px/179% Noto Sans,
-    sans-serif;
-}
-.div-25 {
-  font-family:
-    Noto Sans,
-    sans-serif;
-  margin-top: 64px;
-}
-@media (max-width: 991px) {
-  .div-25 {
-    margin-top: 40px;
-  }
-}
-.div-26 {
-  font-family:
-    Noto Sans,
-    sans-serif;
-  margin-top: 68px;
-}
-@media (max-width: 991px) {
-  .div-26 {
-    margin-top: 40px;
-  }
-}
-.div-27 {
-  display: flex;
-  margin-top: 45px;
-  justify-content: space-between;
-  gap: 20px;
-  font-size: 20px;
-  white-space: nowrap;
-}
-@media (max-width: 991px) {
-  .div-27 {
-    max-width: 100%;
-    flex-wrap: wrap;
-    margin-top: 40px;
-    white-space: initial;
-  }
-}
-.div-28 {
-  color: #333;
-  font-family:
-    Noto Sans,
-    sans-serif;
-  font-weight: 700;
-  line-height: 275%;
-  flex-grow: 1;
-  margin: auto 0;
-}
-@media (max-width: 991px) {
-  .div-28 {
-    white-space: initial;
-  }
-}
-.div-29 {
-  font-family:
-    Noto Sans,
-    sans-serif;
-  border-radius: 5px;
-  border: 1px solid #b9b9b9;
-  background-color: #fff;
-  flex-grow: 1;
-  justify-content: center;
-  align-items: start;
-  color: rgba(255, 0, 0, 0.8);
-  font-weight: 400;
-  line-height: 55px;
-  padding: 18px 60px 18px 20px;
-}
-@media (max-width: 991px) {
-  .div-29 {
-    white-space: initial;
-    max-width: 100%;
-    padding-right: 20px;
-  }
-}
-.div-30 {
-  display: flex;
-  margin-top: 30px;
-  justify-content: space-between;
-  gap: 20px;
-  font-size: 20px;
-  white-space: nowrap;
-  line-height: 275%;
-}
-@media (max-width: 991px) {
-  .div-30 {
-    max-width: 100%;
-    flex-wrap: wrap;
-    white-space: initial;
-  }
-}
-.div-31 {
-  color: #333;
-  font-family:
-    Noto Sans,
-    sans-serif;
-  font-weight: 700;
-  flex-grow: 1;
-  margin: auto 0;
-}
-@media (max-width: 991px) {
-  .div-31 {
-    white-space: initial;
-  }
-}
-.div-32 {
-  font-family:
-    Noto Sans,
-    sans-serif;
-  border-radius: 5px;
-  border: 1px solid #b9b9b9;
-  background-color: #fff;
-  flex-grow: 1;
-  justify-content: center;
-  align-items: start;
-  color: #808080;
-  font-weight: 400;
-  padding: 18px 60px 18px 20px;
-}
-@media (max-width: 991px) {
-  .div-32 {
-    white-space: initial;
-    max-width: 100%;
-    padding-right: 20px;
-  }
-}
-.div-33 {
-  border-radius: 5px;
-  border: 1px solid #b9b9b9;
-  background-color: #fff;
-  align-self: end;
-  margin-top: 30px;
-  width: 550px;
-  max-width: 100%;
-  justify-content: center;
-  align-items: start;
-  color: #808080;
-  white-space: nowrap;
-  padding: 18px 60px 18px 20px;
-  font:
-    400 20px/275% Noto Sans,
-    sans-serif;
-}
-@media (max-width: 991px) {
-  .div-33 {
-    white-space: initial;
-    padding-right: 20px;
-  }
-}
-.div-34 {
-  display: flex;
-  margin-top: 30px;
-  justify-content: space-between;
-  gap: 20px;
-}
-@media (max-width: 991px) {
-  .div-34 {
-    max-width: 100%;
-    flex-wrap: wrap;
-  }
-}
-.div-35 {
-  display: flex;
-  flex-basis: 0%;
-  flex-direction: column;
-  font-size: 20px;
-  color: #333;
-  font-weight: 700;
-  white-space: nowrap;
-  line-height: 275%;
-  margin: auto 0;
-}
-@media (max-width: 991px) {
-  .div-35 {
-    white-space: initial;
-  }
-}
-.div-36 {
-  font-family:
-    Noto Sans,
-    sans-serif;
-}
-.div-37 {
-  font-family:
-    Noto Sans,
-    sans-serif;
-  margin-top: 66px;
-}
-@media (max-width: 991px) {
-  .div-37 {
-    margin-top: 40px;
-    white-space: initial;
-  }
-}
-.div-38 {
-  font-family:
-    Noto Sans,
-    sans-serif;
-  margin-top: 131px;
-}
-@media (max-width: 991px) {
-  .div-38 {
-    margin-top: 40px;
-  }
-}
-.div-39 {
-  font-family:
-    Noto Sans,
-    sans-serif;
-  margin-top: 66px;
-}
-@media (max-width: 991px) {
-  .div-39 {
-    margin-top: 40px;
-  }
-}
-.div-40 {
-  display: flex;
-  flex-grow: 1;
-  flex-basis: 0%;
-  flex-direction: column;
-}
-@media (max-width: 991px) {
-  .div-40 {
-    max-width: 100%;
-  }
-}
-.div-41 {
-  border-radius: 5px;
-  border: 1px solid #b9b9b9;
-  background-color: #fff;
-  justify-content: center;
-  align-items: start;
-  color: #000;
-  white-space: nowrap;
-  padding: 19px 60px 19px 20px;
-  font:
-    400 20px/275% Noto Sans,
-    sans-serif;
-}
-@media (max-width: 991px) {
-  .div-41 {
-    max-width: 100%;
-    padding-right: 20px;
-    white-space: initial;
-  }
-}
-.div-42 {
-  margin-top: 30px;
-}
-@media (max-width: 991px) {
-  .div-42 {
-    max-width: 100%;
-  }
-}
-.div-43 {
-  gap: 20px;
-  display: flex;
-}
-@media (max-width: 991px) {
-  .div-43 {
-    flex-direction: column;
-    align-items: stretch;
-    gap: 0px;
-  }
-}
-.column-3 {
-  display: flex;
-  flex-direction: column;
-  line-height: normal;
-  width: 25%;
-  margin-left: 0px;
-}
-@media (max-width: 991px) {
-  .column-3 {
-    width: 100%;
-  }
-}
-.img-3 {
-  aspect-ratio: 1;
-  object-fit: auto;
-  object-position: center;
-  width: 120px;
-  max-width: 100%;
-}
-@media (max-width: 991px) {
-  .img-3 {
-    margin-top: 10px;
-  }
-}
-.column-4 {
-  display: flex;
-  flex-direction: column;
-  line-height: normal;
-  width: 19%;
-  margin-left: 20px;
-}
-@media (max-width: 991px) {
-  .column-4 {
-    width: 100%;
-  }
-}
-.div-44 {
-  border-radius: 5px;
-  background-color: #f90;
-  justify-content: center;
-  align-self: stretch;
-  color: #fff;
-  white-space: nowrap;
-  text-align: center;
-  width: 100%;
-  margin: auto 0;
-  padding: 19px 24px;
-  font:
-    400 20px/275% Noto Sans,
-    sans-serif;
-}
-@media (max-width: 991px) {
-  .div-44 {
-    margin-top: 40px;
-    white-space: initial;
-    padding: 0 20px;
-  }
-}
-.column-5 {
-  display: flex;
-  flex-direction: column;
-  line-height: normal;
-  width: 57%;
-  margin-left: 20px;
-}
-@media (max-width: 991px) {
-  .column-5 {
-    width: 100%;
-  }
-}
-.div-45 {
-  border-radius: 5px;
-  border: 1px solid #b9b9b9;
-  background-color: #fff;
-  justify-content: center;
-  align-self: stretch;
-  align-items: start;
-  color: #808080;
-  white-space: nowrap;
-  width: 100%;
-  margin: auto 0;
-  padding: 18px 60px 18px 21px;
-  font:
-    400 20px/275% Noto Sans,
-    sans-serif;
-}
-@media (max-width: 991px) {
-  .div-45 {
-    white-space: initial;
-    margin-top: 40px;
-    padding: 0 20px;
-  }
-}
-.div-46 {
-  border-radius: 5px;
-  border: 1px solid #b9b9b9;
-  background-color: #fff;
-  display: flex;
-  margin-top: 30px;
-  justify-content: space-between;
-  gap: 20px;
-  font-size: 20px;
-  color: #000;
-  font-weight: 400;
-  white-space: nowrap;
-  line-height: 275%;
-  padding: 16px 20px;
-}
-@media (max-width: 991px) {
-  .div-46 {
-    max-width: 100%;
-    flex-wrap: wrap;
-    white-space: initial;
-  }
-}
-.div-47 {
-  font-family:
-    Noto Sans,
-    sans-serif;
-  flex-grow: 1;
-  flex-basis: auto;
-  margin: auto 0;
-}
-.img-4 {
-  aspect-ratio: 1.04;
-  object-fit: auto;
-  object-position: center;
-  width: 25px;
-  fill: #ff6700;
-}
-.div-48 {
-  display: flex;
-  margin-top: 30px;
-  width: 100%;
-  justify-content: space-between;
-  gap: 20px;
-  font-size: 20px;
-  font-weight: 400;
-  white-space: nowrap;
-  line-height: 275%;
-}
-@media (max-width: 991px) {
-  .div-48 {
-    max-width: 100%;
-    flex-wrap: wrap;
-    white-space: initial;
-  }
-}
-.div-49 {
-  font-family:
-    Noto Sans,
-    sans-serif;
-  border-radius: 5px;
-  border: 1px solid #b9b9b9;
-  background-color: #fff;
-  flex-grow: 1;
-  justify-content: center;
-  align-items: start;
-  color: #000;
-  padding: 18px 60px 18px 20px;
-}
-@media (max-width: 991px) {
-  .div-49 {
-    padding-right: 20px;
-    white-space: initial;
-  }
-}
-.div-50 {
-  color: #333;
-  font-family:
-    Noto Sans,
-    sans-serif;
-  margin: auto 0;
-}
-.div-51 {
-  border-radius: 5px;
-  border: 1px solid #b9b9b9;
-  background-color: #fff;
-  display: flex;
-  justify-content: space-between;
-  gap: 20px;
-  padding: 22px 20px;
-}
-@media (max-width: 991px) {
-  .div-51 {
-    white-space: initial;
-  }
-}
-.div-52 {
-  color: #000;
-  font-family:
-    Noto Sans,
-    sans-serif;
-}
-.div-53 {
-  color: #808080;
-  font-family:
-    Noto Sans,
-    sans-serif;
-}
-.div-54 {
-  display: flex;
-  margin-top: 44px;
-  width: 100%;
-  justify-content: space-between;
-  gap: 20px;
-  font-size: 20px;
-  color: #333;
-  white-space: nowrap;
-  line-height: 275%;
-}
-@media (max-width: 991px) {
-  .div-54 {
-    max-width: 100%;
-    flex-wrap: wrap;
-    margin-top: 40px;
-    white-space: initial;
-  }
-}
-.div-55 {
-  font-family:
-    Noto Sans,
-    sans-serif;
-  font-weight: 700;
-  flex-grow: 1;
-  margin: auto 0;
-}
-@media (max-width: 991px) {
-  .div-55 {
-    white-space: initial;
-  }
-}
-.div-56 {
-  display: flex;
-  justify-content: space-between;
-  gap: 13px;
-  font-weight: 400;
-}
-@media (max-width: 991px) {
-  .div-56 {
-    max-width: 100%;
-    flex-wrap: wrap;
-    white-space: initial;
-  }
-}
-.div-57 {
-  border-radius: 5px;
-  border: 3px solid #f90;
-  background-color: #fff;
-  align-self: start;
-  width: 28px;
-  height: 26px;
-}
-.div-58 {
-  font-family:
-    Noto Sans,
-    sans-serif;
-  margin: auto 0;
-}
-.div-59 {
-  border-radius: 5px;
-  border: 3px solid #f90;
-  background-color: #f90;
-  aspect-ratio: 1.04;
-  justify-content: center;
-  align-items: center;
-  height: 28px;
-  color: #fff;
-  padding: 0 7px;
-  font:
-    700 25px Sen,
-    sans-serif;
-}
-@media (max-width: 991px) {
-  .div-59 {
-    white-space: initial;
-  }
-}
-.div-60 {
-  font-family:
-    Noto Sans,
-    sans-serif;
-  margin: auto 0;
-}
-.div-61 {
-  border-radius: 5px;
-  border: 3px solid #f90;
-  background-color: #fff;
-  align-self: start;
-  width: 28px;
-  height: 26px;
-}
-.div-62 {
-  font-family:
-    Noto Sans,
-    sans-serif;
-  align-self: start;
-  margin-top: 7px;
-}
-.div-63 {
-  border-radius: 5px;
-  border: 3px solid #f90;
-  background-color: #fff;
-  align-self: start;
-  width: 28px;
-  height: 26px;
-}
-.div-64 {
-  font-family:
-    Noto Sans,
-    sans-serif;
-  margin: auto 0;
-}
-.div-65 {
-  border-radius: 5px;
-  border: 3px solid #f90;
-  background-color: #fff;
-  align-self: start;
-  width: 28px;
-  height: 26px;
-}
-.div-66 {
-  font-family:
-    Noto Sans,
-    sans-serif;
-  margin: auto 0;
-}
-.div-67 {
-  border-radius: 5px;
-  background-color: rgba(255, 153, 0, 0.5);
-  margin-top: 74px;
-  justify-content: center;
-  align-items: center;
-  color: #fff;
-  white-space: nowrap;
-  text-align: center;
-  padding: 16px 60px;
-  font:
-    700 30px/133% Noto Sans,
-    sans-serif;
-}
-@media (max-width: 991px) {
-  .div-67 {
-    max-width: 100%;
-    margin-top: 40px;
-    white-space: initial;
-    padding: 0 20px;
-  }
-}
-.div-68 {
-  color: #777;
-  align-self: center;
-  margin-top: 36px;
-  font:
-    400 20px/150% Noto Sans,
-    sans-serif;
-}
-</style>
